+++ CHANGELOG.md
@@ -25,6 +25,12 @@ Please restart Homebridge after every plugin update.
 * Updated device manufacturer names to reflect official manufacturer names as recorded in OUI docs
 
 
+## 2.1.3 (2023-01-16)
+* Fixed bug: 'Target Media State': characteristic was supplied illegal value #91
+* Fixed bug in setMediaState
+* Improved handling of Target Media State
+
+
 ## 2.1.2 (2023-01-08)
 * Fixed bug causing the plugin to crash when logging an error in getRecordingState and refreshDeviceChannelList